LayoutTests/fast/canvas/webgl/framebuffer-object-attachment-expected.txt

Test framebuffer object attachment behaviors

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Create renderbuffers
PASS gl = create3DContext() is non-null.
PASS colorBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS depthBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS stencilBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS depthStencilBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
Attach depth using DEPTH_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
Attach depth using STENCIL_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
Attach depth using DEPTH_STENCIL_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
Attach stencil using STENCIL_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
Attach stencil using DEPTH_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
Attach stencil using DEPTH_STENCIL_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
Attach depthStencil using DEPTH_STENCIL_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
Attach depthStencil using DEPTH_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
Attach depthStencil using STENCIL_ATTACHMENT
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
PASS gl.getError() is gl.INVALID_FRAMEBUFFER_OPERATION
Attach depth, then stencil, causing conflict
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
Attach stencil, then depth, causing conflict
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
Attach depth, then depthStencil, causing conflict
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
Attach depthStencil, then depth, causing conflict
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
Attach stencil, then depthStencil, causing conflict
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
Attach depthStencil, then stencil, causing conflict
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_UNSUPPORTED
Attach color renderbuffer with internalformat == RGBA4
PASS colorBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
Attach color renderbuffer with internalformat == RGB5_A1
PASS colorBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
Attach color renderbuffer with internalformat == RGB565
PASS colorBuffer = gl.createRenderbuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS fbo = gl.createFramebuffer() is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
